# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Orange Money API",
        "description": "API for Orange Money payment system with US 2.0 compliance",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/v1/comptes": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "List accounts",
                "description": "List all accounts with US 2.0 compliance. Admin sees all accounts, client sees only their own.",
                "operationId": "3963af8778e6c27c99538ec2567d9086",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "maximum": 100
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Account type filter",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "epargne",
                                "cheque"
                            ]
                        }
                    },
                    {
                        "name": "statut",
                        "in": "query",
                        "description": "Account status filter",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "actif",
                                "bloque",
                                "ferme"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by account number or holder name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort field",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "date_ouverture",
                                "solde",
                                "numero_compte",
                                "type",
                                "statut"
                            ]
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Sort order",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "desc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Accounts retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Comptes récupérés avec succès"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Compte"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/v1/comptes/balance": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Get account balance",
                "description": "Get the balance of the user's primary account",
                "operationId": "2b3488c54af9889fdf1db9d13657340b",
                "responses": {
                    "200": {
                        "description": "Balance retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Solde récupéré avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "solde": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 1500.5
                                                },
                                                "solde_formate": {
                                                    "type": "string",
                                                    "example": "1 500,50 FCFA"
                                                },
                                                "numero_compte": {
                                                    "type": "string",
                                                    "example": "OM123456789"
                                                },
                                                "plafond_journalier": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 50000
                                                },
                                                "statut": {
                                                    "type": "string",
                                                    "example": "actif"
                                                },
                                                "devise": {
                                                    "type": "string",
                                                    "example": "FCFA"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Account not found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/v1/comptes/{compte}": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Get account details",
                "description": "Get detailed information about a specific account",
                "operationId": "30e09055c70b19fb0a1e405289cb684d",
                "parameters": [
                    {
                        "name": "compte",
                        "in": "path",
                        "description": "Account ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Account details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Détails du compte récupérés avec succès"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Compte"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Account not found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/v1/comptes/qr-code": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Generate QR code",
                "description": "Generate QR code for the user's primary account",
                "operationId": "62e6be8f0c93b677f9d265fdb476de2b",
                "responses": {
                    "200": {
                        "description": "QR code generated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "QR Code généré avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "qr_code": {
                                                    "description": "Base64 encoded QR code image",
                                                    "type": "string"
                                                },
                                                "numero_compte": {
                                                    "type": "string",
                                                    "example": "OM123456789"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Account not found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nom": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "prenom": {
                        "type": "string",
                        "example": "John"
                    },
                    "nomComplet": {
                        "type": "string",
                        "example": "Doe John"
                    },
                    "telephone": {
                        "type": "string",
                        "example": "771234567"
                    },
                    "email": {
                        "type": "string",
                        "example": "john.doe@example.com"
                    },
                    "statut": {
                        "type": "string",
                        "example": "actif"
                    },
                    "langue": {
                        "type": "string",
                        "example": "fr"
                    },
                    "themeSombre": {
                        "type": "boolean",
                        "example": false
                    },
                    "scannerActif": {
                        "type": "boolean",
                        "example": true
                    },
                    "soldeTotal": {
                        "type": "number",
                        "format": "float",
                        "example": 1500.5
                    },
                    "metadata": {
                        "properties": {
                            "derniereModification": {
                                "type": "string",
                                "format": "date-time"
                            },
                            "dateCreation": {
                                "type": "string",
                                "format": "date-time"
                            },
                            "version": {
                                "type": "integer",
                                "example": 1
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "Compte": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "numero_compte": {
                        "type": "string",
                        "example": "OM123456789"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "epargne",
                            "cheque"
                        ],
                        "example": "epargne"
                    },
                    "solde": {
                        "type": "number",
                        "format": "float",
                        "example": 1500.5
                    },
                    "statut": {
                        "type": "string",
                        "enum": [
                            "actif",
                            "bloque",
                            "ferme"
                        ],
                        "example": "actif"
                    },
                    "plafond_journalier": {
                        "type": "number",
                        "format": "float",
                        "example": 50000
                    },
                    "date_ouverture": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-15"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "client": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            },
            "passport": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}